Make flow run breadcrumbs consistent
